Further break out functions from SwaggerUtil
